--- src/components/TextArea/TextArea2.module.css
@@ -1,33 +1,0 @@
-.component {
-  position: relative;
-  width: 100%;
-  height: 154px;
-  margin-bottom: 10px;
-
-  textarea {
-    width: 100%;
-    height: 154px;
-    padding: 12px;
-    border: none;
-    resize: none;
-
-    &:focus {
-      outline: 2px solid var(--primary);
-      border-radius: var(--rounded-lg);
-    }
-
-    &::-webkit-scrollbar {
-      display: none;
-    }
-  }
-
-  span {
-    position: absolute;
-    bottom: 1px;
-    right: 3px;
-    padding: 5px 12px 12px 5px;
-    color: var(--content-secondary);
-    background-color: #fff;
-    font-size: 0.625rem;
-  }
-}
