+++ src/components/TextArea/style.module.css
@@ -1,12 +1,33 @@
 .component {
-  display: block;
-  margin: 0 auto;
-  border: 1px solid var(--content-secondary);
-  resize: none;
-  padding: var(--spacing-8) var(--spacing-12);
-  border-radius: var(--rounded-lg);
-}
+  position: relative;
+  width: 100%;
+  height: 154px;
+  margin-bottom: 10px;
 
-.component:focus {
-  outline: 1px solid var(--primary);
+  textarea {
+    width: 100%;
+    height: 154px;
+    padding: 12px;
+    border: none;
+    resize: none;
+
+    &:focus {
+      outline: 2px solid var(--primary);
+      border-radius: var(--rounded-lg);
+    }
+
+    &::-webkit-scrollbar {
+      display: none;
+    }
+  }
+
+  span {
+    position: absolute;
+    bottom: 1px;
+    right: 3px;
+    padding: 5px 12px 12px 5px;
+    color: var(--content-secondary);
+    background-color: #fff;
+    font-size: 0.625rem;
+  }
 }
